+++ pkg/graphql/response.go
@@ -1,0 +1,8 @@
+package graphql
+
+import "encoding/json"
+
+type responseBody struct {
+	Data   *json.RawMessage        `json:"data"`
+	Errors ResponseErrorCollection `json:"errors"`
+}