+++ pkg/graphql/spec.go
@@ -1,0 +1,7 @@
+package graphql
+
+import "context"
+
+type Client interface {
+	ExecuteQuery(ctx context.Context, query string, variables map[string]string, v interface{}) error
+}