+++ CHANGELOG.md
@@ -20,6 +20,7 @@ should be used to specify a list of availability zones.
 
 - Build release binaries using go 1.16. Kubectl-gs is now available for Linux and Darwin ARM64 machines including Apple M1 Macs.
 - Upgrade dependency github.com/giantswarm/app to 4.9.0
+- Fetch installation information using a new service, instead of relying on the Giant Swarm REST API.
 
 ## [1.25.0] - 2021-03-16